+++ src/main.ts
@@ -5,6 +5,7 @@ import { NoteDraftRegistry } from "./note-draft-registry";
 import { StandardNotesHost } from "./standard-notes-host";
 import { stampFileProperties } from "./core/file-properties.js";
 import "./styles.css";
+import "./core/icons.css";
 
 declare global {
   interface Window {